# Skip final switch member check if unncessary