@thewilsonator 2021-07-09
```
generated/linux/release/64/benchmark.o:benchmark/runbench.d:_D32TypeInfo_S3std4file11DirIterator6__initZ: error: undefined reference to '_D3std4file11DirIterator9__xtoHashFNbNeKxSQBoQBnQBlZm'
generated/linux/release/64/benchmark.o:benchmark/runbench.d:_D32TypeInfo_S3std4file11DirIterator6__initZ: error: undefined reference to '_D3std4file11DirIterator11__xopEqualsFKxSQBnQBmQBkKxQmZb'
collect2: error: ld returned 1 exit status
```

No idea why. Try force pushing or rebasing?