@Geod24 2021-07-09
Nice! And great explanation. Does it work with enum that have duplicated members ?
E.g.
```D
import std.stdio;

enum Foo
{
 	A = 1,
    B = 1,
    C,
    D,
}

void main ()
{
    Foo f;
    final switch (f)
    {
     	case Foo.A:
        case Foo.C:
        	writeln(f);
	        break;
        case Foo.D:
        	writeln("D: ", f);
        	break;
    }
}
```