@dkorpel 2021-07-09
> Does it work with enum that have duplicated members ?

No, because then members > cases so it has to do the duplicate check.
Ideally the error check is short-circuited by `if (switchCaseCount != uniqueEnumMemberCount)` but dmd currently doesn't keep track of unicity of enum members as far as I see.