@dkorpel 2021-07-11
> Because CircleCI hates @dkorpel apparently. Seems only to happen to them, no idea why. e: its happened on a couple of other PRs

Indeed, CircleCI doesn't even want to start in any of my PRs across dmd/druntime/phobos since 3 days ago. I don't use CircleCI at all so I don't know what the problem could be.